SWITCHYARD-830: Security Policy support: Stage 2
