+++ demos/policy-security-sslbasic/pom.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.switchyard.quickstarts</groupId>
+    <artifactId>switchyard-quickstart-parent</artifactId>
+    <version>0.6.0-SNAPSHOT</version>
+    <relativePath>../../pom.xml</relativePath>
+  </parent>
+
+  <groupId>org.switchyard.quickstarts.demos</groupId>
+  <artifactId>switchyard-quickstart-demo-policy-security-sslbasic</artifactId>
+  <name>Quickstart Demo : Security Policy : SSL + Basic Authentication</name>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.switchyard</groupId>
+      <artifactId>switchyard-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.switchyard</groupId>
+      <artifactId>switchyard-plugin</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.switchyard</groupId>
+      <artifactId>switchyard-test</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.switchyard.components</groupId>
+      <artifactId>switchyard-component-bean</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.switchyard.components</groupId>
+      <artifactId>switchyard-component-soap</artifactId>
+    </dependency>
+  </dependencies>
+  <build>
+    <finalName>${project.artifactId}</finalName>
+
+    <plugins>
+      <plugin>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <configuration>
+          <source>1.6</source>
+          <target>1.6</target>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.switchyard</groupId>
+        <artifactId>switchyard-plugin</artifactId>
+          <executions>
+          <execution>
+            <goals>
+              <goal>configure</goal>
+            </goals>
+            <configuration>
+              <scannerClassNames>
+                <param>org.switchyard.component.bean.config.model.BeanSwitchYardScanner</param>
+                <param>org.switchyard.transform.config.model.TransformSwitchYardScanner</param>
+              </scannerClassNames>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <configuration>
+            <executable>java</executable>
+            <classpathScope>test</classpathScope>
+            <mainClass>org.switchyard.quickstarts.demo.policy.security.sslbasic.WorkServiceMain</mainClass>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+</project>
