+++ demos/policy-security-sslbasic/src/test/resources/xml/soap-request.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<soap:Envelope
+    xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/"
+    xmlns:policy-security-sslbasic="urn:switchyard-quickstart-demo:policy-security-sslbasic:0.1.0">
+    <soap:Body>
+        <policy-security-sslbasic:doWork>
+            <work>
+                <command>WORK_CMD</command>
+            </work>
+        </policy-security-sslbasic:doWork>
+    </soap:Body>
+</soap:Envelope>
