+++ demos/policy-security-sslsaml/picketlink-sts/WEB-INF/classes/picketlink-sts.xml
@@ -1,0 +1,27 @@
+<PicketLinkSTS xmlns="urn:picketlink:identity-federation:config:1.0"
+	STSName="PicketLinkSTS" TokenTimeout="7200" EncryptToken="false">
+	<KeyProvider ClassName="org.picketlink.identity.federation.core.impl.KeyStoreKeyManager">
+		<Auth Key="KeyStoreURL" Value="sts_keystore.jks"/> 
+  		<Auth Key="KeyStorePass" Value="testpass"/>
+  		<Auth Key="SigningKeyAlias" Value="sts"/>
+  		<Auth Key="SigningKeyPass" Value="keypass"/>
+  		<ValidatingAlias Key="http://services.testcorp.org/provider1" Value="service1"/>
+  		<ValidatingAlias Key="urn:switchyard-quickstart-demo:policy-security-sslsaml:0.1.0" Value="service1"/>
+	</KeyProvider>
+	<TokenProviders>
+            <TokenProvider ProviderClass="org.picketlink.identity.federation.core.wstrust.plugins.saml.SAML11TokenProvider"
+                TokenType="http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV1.1"
+	        TokenElement="Assertion"
+	        TokenElementNS="urn:oasis:names:tc:SAML:1.0:assertion"/>
+            <TokenProvider ProviderClass="org.picketlink.identity.federation.core.wstrust.plugins.saml.SAML20TokenProvider"
+                TokenType="http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0"
+	        TokenElement="Assertion"
+	        TokenElementNS="urn:oasis:names:tc:SAML:2.0:assertion"/>
+	</TokenProviders>
+	<ServiceProviders>
+		<ServiceProvider Endpoint="http://services.testcorp.org/provider1" TokenType="http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0"
+			TruststoreAlias="service1"/>
+		<ServiceProvider Endpoint="urn:switchyard-quickstart-demo:policy-security-sslsaml:0.1.0" TokenType="http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0"
+			TruststoreAlias="service1"/>
+	</ServiceProviders>
+</PicketLinkSTS>
