+++ demos/policy-security-sslsaml/picketlink-sts/WEB-INF/classes/roles.properties
@@ -1,0 +1,7 @@
+tomcat=STSClient,JBossAdmin
+JBoss=STSClient
+UserA=STSClient,testRole,JBossAdmin
+UserB=STSClient
+UserC=STSClient
+admin=JBossAdmin
+custom-authenticator-user=STSClient
