+++ demos/policy-security-sslsaml/picketlink-sts/WEB-INF/classes/users.properties
@@ -1,0 +1,6 @@
+JBoss=JBoss
+UserA=PassA
+UserB=PassB
+UserC=PassC
+admin=admin
+tomcat=tomcat
