+++ demos/policy-security-sslsaml/picketlink-sts/WEB-INF/jboss-web.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<jboss-web>
+  <security-domain>picketlink-sts</security-domain>
+  <context-root>picketlink-sts</context-root>
+</jboss-web>
