+++ demos/policy-security-sslsaml/picketlink-sts/WEB-INF/web.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0"?>
+<web-app xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
+	version="2.5">
+	<servlet>
+		<servlet-name>PicketLinkSTS</servlet-name>
+		<servlet-class>org.picketlink.identity.federation.app.sts.PicketLinkSTService</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>PicketLinkSTS</servlet-name>
+		<url-pattern>/*</url-pattern>
+	</servlet-mapping>
+
+	<security-constraint>
+		<web-resource-collection>
+			<web-resource-name>PicketLinkSTSService</web-resource-name>
+			<url-pattern>/*</url-pattern>
+		</web-resource-collection>
+		<auth-constraint>
+			<role-name>JBossAdmin</role-name>
+			<role-name>STSClient</role-name>
+		</auth-constraint>
+	</security-constraint>
+
+	<login-config>
+		<auth-method>BASIC</auth-method>
+		<realm-name>PicketLinkSTSRealm</realm-name>
+	</login-config>
+
+	<security-role>
+		<role-name>JBossAdmin</role-name>
+	</security-role>
+	<security-role>
+		<role-name>STSClient</role-name>
+	</security-role>
+
+</web-app>
