+++ demos/policy-security-sslsaml/src/test/resources/xml/soap-request.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<soap:Envelope
+    xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/"
+    xmlns:wsse="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
+    xmlns:policy-security-sslsaml="urn:switchyard-quickstart-demo:policy-security-sslsaml:0.1.0">
+    <soap:Header>
+        <wsse:Security>
+            <!-- Assertion -->
+        </wsse:Security>
+    </soap:Header>
+    <soap:Body>
+        <policy-security-sslsaml:doWork>
+            <work>
+                <command>WORK_CMD</command>
+            </work>
+        </policy-security-sslsaml:doWork>
+    </soap:Body>
+</soap:Envelope>
