--- demos/policy-security/src/main/java/org/switchyard/quickstarts/demo/policy/security/Work.java
@@ -1,39 +1,0 @@
-/* 
- * JBoss, Home of Professional Open Source 
- * Copyright 2012 Red Hat Inc. and/or its affiliates and other contributors
- * as indicated by the @author tags. All rights reserved. 
- * See the copyright.txt in the distribution for a 
- * full listing of individual contributors.
- *
- * This copyrighted material is made available to anyone wishing to use, 
- * modify, copy, or redistribute it subject to the terms and conditions 
- * of the GNU Lesser General Public License, v. 2.1. 
- * This program is distributed in the hope that it will be useful, but WITHOUT A 
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more detailss. 
- * You should have received a copy of the GNU Lesser General Public License, 
- * v.2.1 along with this distribution; if not, write to the Free Software 
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, 
- * MA  02110-1301, USA.
- */
-package org.switchyard.quickstarts.demo.policy.security;
-
-import java.io.Serializable;
-
-/**
- * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; (C) 2012 Red Hat Inc.
- */
-@SuppressWarnings("serial")
-public class Work implements Serializable {
-
-    private String _command;
-
-    public String getCommand() {
-        return _command;
-    }
-
-    public void setCommand(String command) {
-        _command = command;
-    }
-
-}
