--- demos/policy-security/src/test/resources/xml/soap-request.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<soap:Envelope
-    xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/"
-    xmlns:policy-security="urn:switchyard-quickstart-demo:policy-security:0.1.0">
-    <soap:Body>
-        <policy-security:doWork>
-            <work>
-                <command>WORK_CMD</command>
-            </work>
-        </policy-security:doWork>
-    </soap:Body>
-</soap:Envelope>
